Add SettingsManager integration and enhance Noclip functionality
- Introduced SettingsManager for loading, saving, and resetting settings.
- Updated Main.cpp to load settings on initialization and save before shutdown.
- Enhanced Noclip feature to toggle via hotkey and refined movement speed control.
- Added new UI elements in Menu.cpp for settings management.
- Improved bullet handling in BulletHarvester with a queue system for better performance.
- Adjusted offsets and constants for various entity types in Game and Offsets files.
- Refactored Aimbot and ESP features to utilize new settings and improve accuracy.

+++ Internal DX11 Base/Features/Noclip.h
@@ -51,7 +51,7 @@ private:
     void ApplyMovement(const Utils::Vector3& worldPos);
     uintptr_t GetPositionPointer();
     uintptr_t GetVelocityPointer();
-    void SetVelocity(double velocityX, double velocityY, double velocityZ);
+    void SetVelocity(float velocityX, float velocityY, float velocityZ);
 };
 
 inline Noclip* g_Noclip = nullptr;
